create choukei module
