+++ lib/zakuro/version/gihou/stella/lunar/value.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../../../calculation/type/old_float'
+require_relative '../../../../calculation/stella/lunar/choukei'
 
 require_relative '../../cycle/remainder'
 
@@ -46,8 +46,9 @@ module Zakuro
 
           minus_minute = Adjustment.minus_minute(day: day, minute: minute)
 
-          day = calc_day(per: value.per, denominator: denominator,
-                         minute: minus_minute)
+          day = Calculation::Lunar::Choukei.calc_day(
+            per: value.per, denominator: denominator, minute: minus_minute
+          )
 
           # LOGGER.debug("day: #{day}")
 
@@ -68,39 +69,6 @@ module Zakuro
           raise ArgumentError, "unmatch parameter type: #{remainder.class}"
         end
         private_class_method :valid?
-
-        # :reek:TooManyStatements { max_statements: 6 }
-
-        #
-        # 累計値（大余）を作成する
-        #
-        # @param [Integer] per 増減率
-        # @param [Integer] denominator 小余の分母
-        # @param [Integer] minute 小余
-        #
-        # @return [Integer] 累計値（大余）
-        #
-        def self.calc_day(per:, denominator:, minute:)
-          remainder_minute = Calculation::Type::OldFloat.new((per * minute).to_f)
-          float_day = Calculation::Type::OldFloat.new(remainder_minute.get / denominator)
-          # 切り捨て（プラスマイナスに関わらず小数点以下切り捨て）
-          float_day.floor!
-          day = float_day.get
-          # 繰り上げ結果を足す
-          day += carried_minute(remainder_minute: remainder_minute, denominator: denominator)
-
-          day
-        end
-        private_class_method :calc_day
-
-        def self.carried_minute(remainder_minute:, denominator:)
-          remainder_day = remainder_minute.abs % denominator
-          # 四捨五入（1/2日 以上なら繰り上げる）
-          return remainder_minute.sign if remainder_day >= (denominator / 2)
-
-          0
-        end
-        private_class_method :carried_minute
       end
     end
   end
